+++ problems/maxProfitForStockInSevenDays.go
@@ -1,0 +1,17 @@
+package problems
+
+import (
+	model "probsolbygolang/model"
+)
+
+func FindMaxProfitForStockOverSevenDays(purchaseDates []model.StockPurchaseDates, stockPrice []int) model.StockPurchaseDates {
+	var maxProfit int
+	var dates model.StockPurchaseDates
+	for i := range purchaseDates {
+		profit := stockPrice[purchaseDates[i].Sell] - stockPrice[purchaseDates[i].Buy]
+		if maxProfit < profit {
+			dates = purchaseDates[i]
+		}
+	}
+	return dates
+}