+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Postorder display of BST
 - Inorder display of BST
 - Print the pairs of elements with equal sum in a given array
+- Find the max profit for a stock over seven days
 
 ## Fixed (5th July 2021)
 - Package name issues in model/BST_Node.go